refactor: Move custom hooks from components to hooks
