+++ src/lib/hooks/indicators/top/useEnvelope.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Price } from "../../types/Price";
+import { Price } from "../../../types/Price";
 
 export interface IArgs {
   data: Price[];
