+++ src/lib/hooks/indicators/top/useParabolicSAR.ts
